Extract Print to separate struct
