+++ src/action/mod.rs
@@ -1,23 +1,18 @@
 use nom::{
     branch::alt,
-    bytes::complete::tag,
     character::complete::{multispace0, one_of},
     combinator::map,
-    multi::{many0, separated_list},
+    multi::many0,
     sequence::{delimited, terminated, tuple},
     IResult,
 };
 
-use crate::{
-    basic_types::MutableContext,
-    expression::{parse_expression, Expression},
-    function::Functions,
-    printable::Printable,
-};
+use crate::{basic_types::MutableContext, function::Functions, printable::Printable};
 
 mod assign;
 mod do_while;
 mod if_else;
+mod print;
 mod while_statement;
 
 pub(crate) struct Action {
@@ -55,31 +50,6 @@ trait Statement {
     fn evaluate(&self, functions: &Functions, context: &mut MutableContext) -> Printable<()>;
 }
 
-enum StatementEnum {
-    Print(Vec<Box<dyn Expression>>),
-}
-
-impl Statement for StatementEnum {
-    fn evaluate(&self, functions: &Functions, context: &mut MutableContext) -> Printable<()> {
-        match self {
-            StatementEnum::Print(expressions) => expressions
-                .iter()
-                .fold(Printable::wrap(vec![]), |result, e| {
-                    result.and_then(|mut vec| {
-                        e.evaluate(functions, context).map(|value| {
-                            vec.push(value.coerce_to_string());
-                            vec
-                        })
-                    })
-                })
-                .and_then(|strings| Printable {
-                    value: (),
-                    output: vec![strings.join(" ")],
-                }),
-        }
-    }
-}
-
 fn parse_statements(input: &str) -> IResult<&str, Vec<Box<dyn Statement>>> {
     let parse_single_statement = terminated(
         parse_simple_statement,
@@ -90,20 +60,12 @@ fn parse_statements(input: &str) -> IResult<&str, Vec<Box<dyn Statement>>> {
 
 fn parse_simple_statement(input: &str) -> IResult<&str, Box<dyn Statement>> {
     alt((
-        parse_print_statement,
+        print::parse_print_statement,
         if_else::parse_if_else_statement,
         while_statement::parse_while_statement,
         do_while::parse_do_while_statement,
         assign::parse_assign_statement,
     ))(input)
-}
-
-fn parse_print_statement(input: &str) -> IResult<&str, Box<dyn Statement>> {
-    let parse_separator = delimited(multispace0, one_of(","), multispace0);
-    let parse_expression_list = separated_list(parse_separator, parse_expression);
-
-    let (i, exprs) = delimited(tag("print("), parse_expression_list, one_of(")"))(input)?;
-    Result::Ok((i, Box::new(StatementEnum::Print(exprs))))
 }
 
 #[cfg(test)]
@@ -119,27 +81,12 @@ mod tests {
     }
 
     #[test]
-    fn print_statement_produces_value() {
-        let (functions, mut empty_variables) = empty_functions_and_variables();
-        let mut context = MutableContext::for_variables(&mut empty_variables);
-        context.set_record_with_line("");
-
-        let print_action = parse_action(r#"{ print("hello"); }"#).unwrap().1;
-        assert_eq!(
-            print_action
-                .output_for_line(&functions, &mut context)
-                .output,
-            vec!["hello"],
-        );
-    }
-
-    #[test]
     fn test_parse_statements() {
         let (functions, mut variables) = empty_functions_and_variables();
         let mut context = MutableContext::for_variables(&mut variables);
         context.set_record_with_line("");
 
-        let result = parse_print_statement(r#"print("hello")"#);
+        let result = print::parse_print_statement(r#"print("hello")"#);
         assert!(result.is_ok());
         assert_eq!(
             Action {
