+++ src/action/print.rs
@@ -1,0 +1,75 @@
+use nom::{
+    bytes::complete::tag,
+    character::complete::{multispace0, one_of},
+    multi::separated_list,
+    sequence::delimited,
+    IResult,
+};
+
+use super::Statement;
+use crate::{
+    basic_types::MutableContext,
+    expression::{parse_expression, Expression},
+    function::Functions,
+    printable::Printable,
+};
+
+struct Print {
+    expressions: Vec<Box<dyn Expression>>,
+}
+
+impl Statement for Print {
+    fn evaluate(&self, functions: &Functions, context: &mut MutableContext) -> Printable<()> {
+        self.expressions
+            .iter()
+            .fold(Printable::wrap(vec![]), |result, e| {
+                result.and_then(|mut vec| {
+                    e.evaluate(functions, context).map(|value| {
+                        vec.push(value.coerce_to_string());
+                        vec
+                    })
+                })
+            })
+            .and_then(|strings| Printable {
+                value: (),
+                output: vec![strings.join(" ")],
+            })
+    }
+}
+
+pub(super) fn parse_print_statement(input: &str) -> IResult<&str, Box<dyn Statement>> {
+    let parse_separator = delimited(multispace0, one_of(","), multispace0);
+    let parse_expression_list = separated_list(parse_separator, parse_expression);
+
+    let (i, exprs) = delimited(tag("print("), parse_expression_list, one_of(")"))(input)?;
+    Result::Ok((i, Box::new(Print { expressions: exprs })))
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::action::parse_action;
+    use crate::basic_types::Variables;
+    use crate::function::Functions;
+    use std::collections::HashMap;
+
+    fn empty_functions_and_variables() -> (Functions, Variables) {
+        let variables = Variables::empty();
+        (HashMap::new(), variables)
+    }
+
+    #[test]
+    fn print_statement_produces_value() {
+        let (functions, mut empty_variables) = empty_functions_and_variables();
+        let mut context = MutableContext::for_variables(&mut empty_variables);
+        context.set_record_with_line("");
+
+        let print_action = parse_action(r#"{ print("hello"); }"#).unwrap().1;
+        assert_eq!(
+            print_action
+                .output_for_line(&functions, &mut context)
+                .output,
+            vec!["hello"],
+        );
+    }
+}
